third_party/blink/web_tests/fast/forms/form-radio-img-node-list-expected.txt

Test RadioNodeLists returned by the HTMLFormElement named-getter.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Check that if no 'listed elements' match by name, img elements are picked instead.
PASS radioNodeList.length is 2
PASS radioNodeList[0] instanceof HTMLImageElement is true
PASS radioNodeList[1] instanceof HTMLImageElement is true
PASS radioNodeList.length is 2
PASS radioNodeList[0] instanceof HTMLImageElement is true
PASS radioNodeList[1] instanceof HTMLImageElement is true
PASS radioNodeList.length is 3
PASS radioNodeList[0] instanceof HTMLImageElement is true
PASS radioNodeList[1] instanceof HTMLImageElement is true
PASS radioNodeList[2] instanceof HTMLImageElement is true
PASS successfullyParsed is true

TEST COMPLETE
